fix: itemsテーブルのprice型をintegerに修正 + HasFactory追加

+++ api/database/migrations/2026_04_04_045434_create_items_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->string('name');
             $table->string('brand')->nullable();
-            $table->tinyInteger('price')->unsigned();
+            $table->integer('price')->unsigned();
             $table->text('description')->nullable();
             $table->string('image_url')->nullable();
             $table->tinyInteger('condition')->nullable();
